Refactor module loading and enhance configuration management in counterfactual and solace scripts. Added support for mixed precision configurations and improved error handling. Updated evaluation and reranking scripts with new metrics and logging capabilities. Enhanced shell scripts for better environment variable management.

+++ solace/counterfactual_sd3_utils.py
@@ -62,7 +62,7 @@ def load_sd3_pipeline(config, device, is_local_main_process: bool = True):
     pipeline.text_encoder.to(device, dtype=inference_dtype)
     pipeline.text_encoder_2.to(device, dtype=inference_dtype)
     pipeline.text_encoder_3.to(device, dtype=inference_dtype)
-    pipeline.transformer.to(device)
+    pipeline.transformer.to(device, dtype=inference_dtype)
 
     text_encoders = [pipeline.text_encoder, pipeline.text_encoder_2, pipeline.text_encoder_3]
     tokenizers = [pipeline.tokenizer, pipeline.tokenizer_2, pipeline.tokenizer_3]
